Return `MISSING_GINI_CAPTURE_INSTANCE` error
When the `GiniCapture` instance is missing then the fragment
listeners' `onError()` method is called with the new error.

The `MultiPageReviewFragmentListener` did not have an `onError()`
method so we added one and the `MultiPageReviewActivity` returns the
error as the activity result.

In the Screen API this new error is returned as the activity result.

Activities which don't have listeners simply finish, if there is no
`GiniCapture` instance as these are all started from the camera
screen.