Remove items from the `ParcelableMemoryCache` with a delay
Cloud Messaging version 22.0.0 introduced an activitiy lifecycle
callback listener called `FcmLifecycleCallbacks`. This reads a
cloud messaging analytics related intent extra on activity creations
which causes all parcelables to be unparcelled.

When not using the multi-page feature we pass the documents between
the activites and fragments as parcelables and we use the
`ParcelableMemoryCache` to preserve a reference to the byte arrays
until the document is unparcelled again.

Because `FcmLifecycleCallbacks` caused an unexpected unparcelling the
cached byte array was removed prematurely. When the review activity
was created after taking a picture and it unparcelled the document
then the byte array was already gone.

By introducing a delay for removal we make sure that any unparcelling
between activity switching does not prematurely remove the cached
byte arrays.