Allow showing custom help items
By passing a list of `HelpItem.Custom` objects to `setCustomHelpItems()`
when creating a new instance of `GiniCapture` we can show custom help
items.

Converted `HelpItem` from an enum to a kotlin sealed class to take
advantage of kotlin's `when` statement for dealing with the different
help item types.

Also converted `HelpItemsAdapter` and `HelpActivity` to kotlin.